Separate "main" from client code
